refactoring block
