+++ js/main.js
@@ -1,2 +1,5 @@
-import {createPhotoData} from './data.js';
-createPhotoData();
+import {
+  createPhotoData,
+  NUMBER_OF_OBJECTS_GENERATED
+} from './create-foto-data.js';
+createPhotoData(NUMBER_OF_OBJECTS_GENERATED);
